+++ apps/ledger-live-desktop/src/config/urls.ts
@@ -201,6 +201,9 @@ export const urls = {
   memoTag: {
     learnMore: "https://support.ledger.com/article/4409603715217-zd",
   },
+  geoBlock: {
+    learnMore: "https://support.ledger.com/article/Why-Ledger-Complies-with-Sanctions",
+  },
 };
 
 export const vaultSigner = {