+++ apps/ledger-live-desktop/tests/handlers/index.ts
@@ -1,7 +1,8 @@
 import NFTsHandlers from "./nfts";
 import MarketHandlers from "./market";
+import CryptoIconsHandlers from "./cryptoIcons";
 
-export default [...NFTsHandlers, ...MarketHandlers];
+export default [...NFTsHandlers, ...MarketHandlers, ...CryptoIconsHandlers];
 
 export const ALLOWED_UNHANDLED_REQUESTS = [
   "https://nft.api.live.ledger.com/v2/marketdata/ethereum/1/contract/0xe3BE0054Da2F8da5002E8bdD8AA4c7fDf851E86D/floor-price",