+++ apps/ledger-live-mobile/src/components/RootNavigator/types/SendFundsNavigator.ts
@@ -205,19 +205,6 @@ export type SendFundsNavigatorStackParamList = {
       | ScreenName.SendSelectDevice
       | ScreenName.SwapForm;
   };
-  [ScreenName.XrpEditFee]: {
-    accountId: string;
-    parentId?: string;
-    transaction: RippleTransaction;
-    currentNavigation:
-      | ScreenName.SignTransactionSummary
-      | ScreenName.SendSummary
-      | ScreenName.SwapForm;
-    nextNavigation:
-      | ScreenName.SignTransactionSelectDevice
-      | ScreenName.SendSelectDevice
-      | ScreenName.SwapForm;
-  };
   [ScreenName.StellarEditCustomFees]: {
     accountId: string;
     parentId?: string;