+++ e2e/mobile/page/accounts/addAccount.drawer.ts
@@ -1,6 +1,5 @@
 import { openDeeplink } from "../../helpers/commonHelpers";
 import CommonPage from "../common.page";
-import { getEnv } from "@ledgerhq/live-env";
 import { retryUntilTimeout } from "../../utils/retry";
 
 export default class AddAccountDrawer extends CommonPage {
@@ -12,8 +11,7 @@ export default class AddAccountDrawer extends CommonPage {
   addFundsButtonId = "button-add-funds";
   actionDrawerReceiveButtonId = "action-drawer-receive-button";
 
-  accountId = (currency: string, index: number) =>
-    getEnv("MOCK") ? `mock:1:${currency}:MOCK_${currency}_${index}:` : `js:2:${currency}:.*`;
+  accountIdAccountDrawer = (currency: string) => `js:2:${currency}:.*`;
 
   @Step("Open add account via deeplink")
   async openViaDeeplink() {
@@ -42,7 +40,7 @@ export default class AddAccountDrawer extends CommonPage {
   @Step("Expect account discovered")
   async expectAccountDiscovery(currencyName: string, currencyId: string, index = 0) {
     const accountName = `${currencyName} ${index + 1}`;
-    await detoxExpect(this.accountItem(this.accountId(currencyId, index))).toBeVisible();
+    await detoxExpect(this.accountItem(this.accountIdAccountDrawer(currencyId))).toBeVisible();
     const accountId = (await getIdByRegexp(this.accountItemRegExp(), index)).replace(
       this.accountItemId,
       "",