+++ libs/env/src/env.ts
@@ -87,16 +87,6 @@ const envDefinitions = {
     parser: boolParser,
     desc: "Enable tokens on Aptos",
   },
-  API_CASPER_INDEXER_ENDPOINT: {
-    parser: stringParser,
-    def: "https://casper.coin.ledger.com/indexer",
-    desc: "Casper API url indexer",
-  },
-  API_CASPER_NODE_ENDPOINT: {
-    parser: stringParser,
-    def: "https://casper.coin.ledger.com/node/",
-    desc: "Casper API url node",
-  },
   API_ALGORAND_BLOCKCHAIN_EXPLORER_API_ENDPOINT: {
     def: "https://algorand.coin.ledger.com",
     parser: stringParser,