+++ libs/ledgerjs/jest.config.ts
@@ -6,5 +6,16 @@ export default {
   testPathIgnorePatterns: ["packages/*/lib-es", "packages/*/lib"],
   coveragePathIgnorePatterns: ["packages/create-dapp"],
   passWithNoTests: true,
+  coverageReporters: ["json", ["lcov", { projectRoot: "../../../../" }], "json-summary", "text"],
+  reporters: [
+    "default",
+    [
+      "jest-sonar",
+      {
+        outputName: "sonar-executionTests-report.xml",
+        reportedFilePath: "absolute",
+      },
+    ],
+  ],
   // collectCoverageFrom: ["packages/**/src/*.ts"],
 };